Separate utils into header and cpp, rename sin() and cos() to avoid confusion.
